doc: Add note on password in installation instructions
When the password has special characters that are reserved by yaml, it
must be quoted so it is parsed and extracted correctly. Although it
seems obvious, not everyone is aware of this and the way this issue
manifests is rather cryptic (says connection refused).

Fix this by quoting the password field and adding a note about it.

Signed-off-by: Siddharth Chandrasekaran <sidcha.dev@gmail.com>